Extract decos-service to its own service
