+++ src/server/services/toeristische-verhuur/toeristische-verhuur-notifications.ts
@@ -14,7 +14,7 @@ import { isRecentNotification } from '../../../universal/helpers/utils';
 import { MyNotification } from '../../../universal/types';
 import { AuthProfileAndToken } from '../../auth/auth-types';
 import { NOTIFICATION_REMINDER_FROM_MONTHS_NEAR_END } from '../vergunningen-v2/config-and-types';
-import { isNearEndDate } from '../vergunningen-v2/helpers';
+import { isNearEndDate } from '../decos/helpers';
 
 export function createToeristischeVerhuurNotification(
   vergunning: ToeristischeVerhuurVergunning,
