+++ src/server/services/vergunningen-v2/decos-zaken.ts
@@ -41,7 +41,7 @@ import {
   getCustomTitleForVergunningWithLicensePlates,
   transformBoolean,
   transformKenteken,
-} from './helpers';
+} from '../decos/helpers';
 import {
   caseNotificationLabelsDefault,
   caseNotificationLabelsExpirables,
